Widgets module is ready for public BETA. Time to merge with Master branch ?

+++ application/modules/widgets/models/widgets_m.php
@@ -8,98 +8,13 @@
  * Widgets model that handles database related actions for the widget controller.
  */
 class Widgets_m extends Model
-{		
-	// THE FOLLOWING FUNCTIONS ARE FOR THE FRONTEND ONLY !
-	
-	// Execute all widgets for the specified area
-	function area($area)
-	{
-		// First we need to fetch all activated widgets for the chosen area
-		$this->db->select(array('name','area'));
-		$query = $this->db->get_where('widgets',array('area' => $area));
-		
-		// Verify the results
-		if($query->num_rows() > 0)
-		{
-			// Store the results
-			$widgets = $query->result();
-			
-			// Display the widgets in an unordered list. TODO: Check whether this can't be done in a better way, this is kinda ugly...
-			echo '<ul class="widgets_list">';
-			
-			// Execute each widget
-			foreach($widgets as $widget)
-			{				
-				// Only run the widgets for the current area
-				if($widget->area == $area)
-				{
-					// Validate whether the widget really exists
-					if(file_exists(APPPATH . "widgets/$name/$name.php") == TRUE)
-					{
-						// Open the widget
-						require_once(APPPATH . "widgets/$name/$name.php");
-
-						// First letter needs to be uppercase, in case the user forgets this
-						$class  = ucfirst($name);
-
-						// Create the widget class
-						$widget = new $class();			
-
-						// Verify if we can actually call the function at all
-						if(is_object($widget))
-						{		
-							// Execute the run() function with the specified arguements
-							return $widget->run();
-						}
-						else
-						{
-							// Log the error
-							log_message('error','The run function of the ' . $widget->name . ' widget could not be executed.');
-							return FALSE;
-						}
-					}
-					else
-					{
-						// Log the error
-						log_message('error','The ' . $widget->name . ' widget could not be found. Are you sure it exists ?');
-						return FALSE;
-					}
-				}			
-			}
-			
-			// End of the unordered list
-			echo '</ul>';
-		}
-		else
-		{
-			return FALSE;
-		}
-	}
-	
-	// Display the widget's view files
-	public function display($name,$view,$data = array())
-	{		
-		// Verify the view file
-		if(file_exists(APPPATH . "widgets/$name/views/$view.php"))
-		{
-			// Load the view file
-			$view_file = $this->load->view('../widgets/'.$name.'/views/'.$view, $data,true);
-			echo "<li class='widget $name'>$view_file</li>";
-		}
-		else
-		{		
-			log_message('error','The ' . $view . ' view for the ' . $name . ' widget could not be found.');
-			return FALSE;
-		}		
-	}
-	
-	
-	// THE FOLLOWING FUNTIONS ARE FOR THE BACKEND ONLY !
-	
+{	
 	// Get a list of widgets based on the parameters defined by the user.
 	function getWidgets($params = array())
 	{
 		// Create the query
+		$this->db->select('widget_instances.id,widgets.name,widget_instances.body,widget_instances.area');
+		$this->db->join('widget_instances','widgets.id = widget_instances.widget_id');
 		$query = $this->db->get_where('widgets',$params);
 		
 		// Validate the results
@@ -119,8 +34,8 @@ class Widgets_m extends Model
 		// Terminate the function if the ID isn't specified or when it isn't numeric.
 		if(isset($id) AND is_numeric($id))
 		{
-			$this->db->where('id',$id);
-			$query = $this->db->update('widgets',$params);
+			$this->db->where('widget_instances.id',$id);
+			$query = $this->db->update('widget_instances',$params);
 			
 			return $query;
 		}
@@ -130,17 +45,28 @@ class Widgets_m extends Model
 		}
 	}
 	
-	// Function to delete a widget from the database table.
+	// Function to delete a widget from the database table. This will remove both the instance and regular row.
 	function removeWidget($id)
 	{		
 		// Valid ID ? 
 		if(isset($id) AND is_numeric($id))
 		{
 			// Remove the widget from the database			
-			$this->db->where('id',$id);
-			$query = $this->db->delete('widgets');
+			$this->db->where('widget_id',$id);
+			$query_1 = $this->db->delete('widget_instances');
 			
-			return $query;
+			// Instance deleted, continue
+			if($query_1 == TRUE)
+			{
+				$this->db->where('widgets',$id);
+				$query = $this->db->delete('widgets');
+				
+				return $query;
+			}
+			else
+			{
+				return FALSE;
+			}
 		}
 		else
 		{
@@ -148,22 +74,76 @@ class Widgets_m extends Model
 		}
 	}
 	
+	// Function to get a list of widgets that can be installed
+	function getInstallableWidgets()
+	{
+		// Variables
+		$path 				= APPPATH . 'widgets/';
+		$handle 			= opendir($path);
+		$widget_xml 		= array();
+		$installed_widgets 	= array();
+		
+		// Get a list of widgets that are already installed 
+		$this->db->select('name');
+		$query = $this->db->get('widgets');
+		
+		// Store the names of all widgets currently installed
+		if($query->num_rows() > 0)
+		{
+			$query_results = $query->result();
+			
+			foreach($query_results as $installed_widget)
+			{
+				$installed_widgets[] = $installed_widget->name;
+			}
+		}
+		
+		foreach(glob($path . '*',GLOB_ONLYDIR) as $dir)
+		{
+			if($dir != 'application/widgets/dummy_widget')
+			{
+				$widget_name = str_replace('application/widgets/','',$dir);
+				
+				// Only add the number to the array if the folder contains a file with the same name as the directory and a .php extension (otherwise the system might be fooled).
+				if(file_exists($dir . '/' . $widget_name . '.php') AND file_exists($dir . '/widget.xml'))
+				{
+					if(in_array($widget_name,$installed_widgets) == FALSE)
+					{
+						$widget_xml[] = $this->widgets->parse_xml($widget_name);
+					}
+				}			
+				else
+				{
+					return FALSE;
+				}
+			}			
+		}	
+		// Return results
+		return $widget_xml;		
+	}
+	
 	// Installing widgets. Function should only be used when the widget is activated for the first time.
 	function installWidget($name)
 	{
-		if(isset($name) AND $name != '')
+		// First we need the XML data of the widget
+		$widget_data  = $this->widgets->parse_xml($name);
+		$widget_data  = $widget_data[0];
+		
+		// Insert the name into the database table "widgets"
+		$query 		  = $this->db->insert('widgets',array('name' => $widget_data->filename));
+		
+		if($query != FALSE)
 		{
-			// First we need to retrieve the data for the current widget
-			// TODO: Add this stuff
+			$insert_id 	= $this->db->insert_id();
+
+			// Insert the data into the table "widget_instances"
+			$body 		= serialize($widget_data->body);
+			$query 		= $this->db->insert('widget_instances',array('widget_id' => $insert_id,'body' => $body));	
 			
-			// Now we can install it
-			$install_data['name'] = $name;
-			$install_data['body'] = '';
+			return $query;
 		}
-		else
-		{
-			return FALSE;
-		}
+		
+		return $query;
 	}	
 }
 ?>
